+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/treeitems/HeaderItem.java
@@ -45,6 +45,6 @@ public class HeaderItem implements TreeItem {
 
     @Override
     public boolean isModified() {
-        return false; //??
+        return swf.isHeaderModified();
     }
 }