+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Better tag error handling - these tags now got error icon
 - Show in Hex dump command from other views for tags
 - Show in Taglist command from dump view for tags
+- Create new empty SWF file
 
 ### Fixed
 - Flash viewer - subtract blend mode
@@ -23,6 +24,7 @@ All notable changes to this project will be documented in this file.
 - [#1785] AS1/2 try..catch block in for..in
 - [#1770] Links in basictag info (like needed/dependent characters) were barely visible on most themes
 - Show in Resource command from Hex dump not working for tags inside DefineSprite
+- File did not appear modified when only header was modified
 
 ### Changed
 - [#1455] All tag types are now allowed inside DefineSprite